Change some types

+++ lib/filter/src/Filter.h
@@ -7,8 +7,8 @@
 class Filter {
     public:
         virtual void reset() = 0;
-        virtual void addValue(float newValue) = 0;
-        virtual float getFilteredValue() = 0;
+        virtual void addValue(double newValue) = 0;
+        virtual double getFilteredValue() = 0;
         virtual ~Filter() {}
 };
